Refactor bandplans to reduce gtk dependency
We need the BandPlans object in both UIs, but only the select_bandplan
method depends on GTK. So split those apart so we can re-use the
generic part in wxui.
